Create [Minecraft Series #1] Steve wants to build a beacon pyramid (JavaScript)
Just another day in the world of Minecraft, Steve is working on his new project -- building a beacon pyramid in front of his house.

Steve has already obtained the beacon (the glass wrapped blue thingy on the top), he just needs to build the pyramid. Each level of the pyramid is built using one of the following four kinds of block: gold, diamond, emerald, and iron. Since these four kinds of blocks are relatively hard to collect, Steve needs to know exactly how many of each kind is required to build a pyramid of level N.
Assume that the top level of the pyramid uses gold blocks, the second level uses diamond blocks, the third level uses emerald blocks, the fourth level uses iron blocks, and the fifth level uses gold blocks, etc. (if you are a Minecraft player, you know it's not neccessary the right way to build the pyramid. Let's just assume it is for the sake of this kata ;-))
In JavaScript:
write a function blocksToCollect that takes one argument which is the level of the pyramid, and returns an object of the form: {total: 9, gold: 9, diamond: 0, emerald: 0, iron: 0}